Add battery
Add `status_battery` to check capacity% of the battery.
From now on I have multiple functions which push a string to the
string `status` via `snprintf`, I need to check the length of each
output string (output of `snprintf`) and update the location of `status`
to be updated.
Find a unicode (extended) character to display the battery icon.